requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
psycopg2==2.9.4
base58==2.1.1
billiard==3.6.4.0
bitarray==2.7.2
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.12.1
Django==3.2.18
importlib-metadata==4.13.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-redis==5.2.0
django-timezone-field==5.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
kombu==5.2.4
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
parsimonious==0.8.1
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==3.19.5
pycryptodome==3.17
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
rlp==2.0.1
rq==1.13.0
six==1.16.0
dj-database-url==1.2.0
sqlparse==0.4.3
toolz==0.12.0
tzdata==2022.7
urllib3==1.26.14
varint==1.0.2
vine==5.0.0
wcwidth==0.2.6
web3==5.31.3
websockets==9.1
whitenoise==6.4.0
yarl==1.8.2
sqlalchemy==2.0.0